imporve eval
